# academify/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Academify API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Student Community & Collaboration Platform — COMP6703001"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://e2526-wads-b4ac-02.csbihub.id",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "HTTP-only session cookie set by Better Auth after sign-in. Sign in via the app or POST /api/auth/sign-in/email, then use Try it out from the same origin."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Legacy alias — Academify uses session cookies, not Bearer JWT."
      }
    },
...
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "summary": "Get all forum threads",
        "tags": [
          "Forums"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "forumId",
            "schema": {
              "type": "string"
            },
            "description": "Filter threads by forum ID"
          },
          {
            "in": "query",
            "name": "forum",
            "schema": {
              "type": "string"
            },
            "description": "Filter threads by forum name or slug"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of results to return"
          },
          {
            "in": "query",
            "name": "trending",
            "schema": {
              "type": "boolean"
            },
            "description": "Return trending threads"
          }
        ],
        "responses": {
          "200": {
            "description": "List of threads"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new thread",
        "tags": [
          "Forums"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content",
                  "forumId"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Best resources for learning React?"
                  },
                  "content": {
                    "type": "string",
                    "example": "Looking for good React learning resources."
                  },
                  "forumId": {
                    "type": "string",
                    "example": "forum_cuid"
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content",
                  "forum"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "forum": {
                    "type": "string",
                    "description": "Forum name or slug"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thread created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/reports/{reportId}/action": {
      "post": {
        "summary": "Take action on a report (Moderator/Admin only)",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reportId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "resolve",
                      "dismiss"
                    ],
                    "example": "resolve"
                  },
                  "note": {
                    "type": "string",
                    "example": "Content removed and user warned."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Action taken on report"
          },
          "400": {
            "description": "Missing or invalid action"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden - Moderator/Admin only"
          },
          "404": {
            "description": "Report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}